More accessibility and yoga renamings
