+++ ReactCommon/fabric/components/view/ConcreteViewShadowNode.h
@@ -10,7 +10,7 @@
 #include <fabric/components/view/accessibility/AccessibleShadowNode.h>
 #include <fabric/components/view/ViewEventEmitter.h>
 #include <fabric/components/view/ViewProps.h>
-#include <fabric/components/view/YogaLayoutableShadowNode.h>
+#include <fabric/components/view/yoga/YogaLayoutableShadowNode.h>
 #include <fabric/core/ConcreteShadowNode.h>
 #include <fabric/core/LayoutableShadowNode.h>
 #include <fabric/core/ShadowNode.h>
